chore: move password util
