+++ src/config/passport.js
@@ -6,7 +6,7 @@ const ExtractJWT = passportJWT.ExtractJwt;
 const prisma = require('../config/prisma');
 const GeneralError = require('../errors/error/general-error');
 const signInErrorMap = require('../errors/sign-in-error');
-const { comparePassword } = require('../utils/func/password');
+const { comparePassword } = require('../utils/crypto/password');
 const { jwtSecret } = require('../config/config');
 
 passport.use(
